app/requirements.txt
Django>=4.2,<5.0
gunicorn
psycopg2-binary
django-environ
djangorestframework
djangorestframework-simplejwt
Pillow